refactor: 함수명과 파일명  수정 및 모듈 분리
함수명과 파일명을 수정하고 모듈을 분리한다
