+++ READEME/week3.md
@@ -10,3 +10,6 @@
 
 
 ### **🤔 생각해볼 거리**
+
+문제점 : fetch 한 데이터를 렌더링 시킬 때 비동기로 데이터를 가져와 함수를 실행시키기 때문에
+함수를 호출한 순서대로 데이터가 렌더링 되지 않고, fetch한 순서대로 데이터를 랜더링 한다
