--- src/js/carousel/init-carousel-event.js
@@ -1,15 +1,0 @@
-import {
-  slideLoadEventHandler,
-  slideNextBtnClickEventHandler,
-  slidePrevBtnClickEventHandler,
-  slideTransitionendEventHandler,
-  slideAutoEventHandler,
-} from "./carousel-event.js";
-
-export function initSlideEvent() {
-  slideLoadEventHandler();
-  slideNextBtnClickEventHandler();
-  slidePrevBtnClickEventHandler();
-  slideTransitionendEventHandler();
-  slideAutoEventHandler();
-}
