+++ src/js/util/delay.js
